Extract shell local command executor
